Skip the starting waypoint
If there is a starting point, it produces a waypoint somewhere in the path (either the first or the last point). This waypoint shouldn't be included in the eventual path. Skip it. This can be done a bit more efficient but that is of later concern.

Contributes to issue CURA-329.