Remove duplicate code for reverse paths
In the previous commit I immediately finalised reverse paths in their waypoints as soon as it was known that the path was to be reversed. This changed the code for the non-reversed try (when reverse paths are allowed) more or less equal to the code for when reversing paths isn't allowed. I removed the code for this non-reversed try, making it always try non-reversed and only try reversed if reversed is allowed.

Contributes to issue CURA-329.