Store polyStart during clustering
During clustering, the best orientation of the line is already decided in order to find the closest line. This orientation is immediately stored at that time, instead of guessing it later by the order of the lines at the serialisation. This should make the algorithm more robust at no additional computational cost.

Contributes to issue CURA-329.