Order with average points and insert starting point last
This commit has two major changes for which a rewrite was required: The order of the pieces is now determined with point-based TSP rather than line-based. The average between the start and end point is used. This provides better results if the clusters are big, but worse results if the lines intersect (which is not our use case). Then to determine the direction of the elements, it is simply iterated through one by one after all elements are in the sequence.
Also, rather than inserting the starting point at the start, it is now inserted at the end. If it turns out that the starting point should've been inserted somewhere in the middle, that will be corrected after all elements are inserted.

Contributes to issue CURA-329.