Improve clustering heuristic when not aligned
In grid infill, the infill lines are not necessarily aligned. This makes clustering more difficult, but this should help at no additional computational cost. The best distance is now only computed from the travel move that's to be made, while the check for both endpoints to be near remains in effect.

Contributes to issue CURA-329.