Outline the procedures to insert waypoints
In the optimisation step, outline the procedures to insert elements in the resulting lists to their own functions. I really don't agree with this change, since I think it reduces readability as well as efficiency, but it was desired by code review to reduce the deepest indentation of the code, which is a good thing.

Contributes to issue CURA-329.