Alternative implementation of BucketGrid2D
CURA-590

Instead of mapping a cell location multiple times to an object within the cell, each cell location is mapped only once to a vector of objects within the cell.
The first one has the overhead of 'bucket-collisions' where all mappings of two different cells get placed in the same bucket, which causes findNearby to loop over unneeded elements.
The second has the overhead and indirection of creating vectors and all that comes with it.